+++ src/component/Schema/Params/Page/assertContent.js
@@ -1,6 +1,6 @@
 import { buildAssertionTpl } from "service/assert";
 import { AssertValue } from "../../Assert/AssertValue";
-import { truncate } from "service/utils";
+import { truncate, normalizeAssertionVerb } from "service/utils";
 
 export const assertContent = {
   template: ( command ) => buildAssertionTpl(
@@ -11,8 +11,8 @@ export const assertContent = {
   description: `Asserts that the page content (HTML) satisfies the given constraint`,
   commonly: "assert page HTML",
 
-  toLabel: ({ assert }) => `(${ assert.assertion } \`${ truncate( assert.value, 60 ) }\`)`,
-  toText: ({ assert }) => `(${ assert.assertion } \`${ assert.value }\`)`,
+  toLabel: ({ assert }) => `(${ normalizeAssertionVerb( assert.assertion ) } \`${ truncate( assert.value, 60 ) }\`)`,
+  toText: ({ assert }) => `(${ normalizeAssertionVerb( assert.assertion ) } \`${ assert.value }\`)`,
 
   assert: {
     node: AssertValue