+++ src/component/Schema/Params/Page/assertUrl.js
@@ -1,6 +1,6 @@
 import { buildAssertionTpl } from "service/assert";
 import { AssertValue } from "../../Assert/AssertValue";
-import { truncate } from "service/utils";
+import { truncate, normalizeAssertionVerb } from "service/utils";
 
 export const assertUrl = {
   template: ( command ) => buildAssertionTpl(
@@ -9,8 +9,8 @@ export const assertUrl = {
     `// Asserting that page URL satisfies the given constraint`
   ),
 
-  toLabel: ({ assert }) => `(${ assert.assertion } \`${ truncate( assert.value, 60 ) }\`)`,
-  toText: ({ assert }) => `(${ assert.assertion } \`${  assert.value, 60 }\`)`,
+  toLabel: ({ assert }) => `(${ normalizeAssertionVerb( assert.assertion ) } \`${ truncate( assert.value, 60 ) }\`)`,
+  toText: ({ assert }) => `(${ normalizeAssertionVerb( assert.assertion ) } \`${  assert.value, 60 }\`)`,
   commonly: "assert page URL",
 
   description: `Asserts that the page URL satisfies the given constraint`,