+++ src/component/Schema/Params/Page/assertVar.js
@@ -1,7 +1,7 @@
 import { buildAssertionTpl } from "service/assert";
 import { AssertValue } from "../../Assert/AssertValue";
 import { INPUT } from "../../constants";
-import { truncate } from "service/utils";
+import { truncate, normalizeAssertionVerb } from "service/utils";
 
 export const assertVar = {
   template: ( command ) => buildAssertionTpl(
@@ -10,8 +10,8 @@ export const assertVar = {
     `// Asserting that variable associated with a given name satisfies the given constraint`
   ),
 
-  toLabel: ({ assert }) => `(${ assert.assertion } \`${ truncate( assert.value, 60 ) }\`)`,
-  toText: ({ assert }) => `(${ assert.assertion } \`${  assert.value, 60 }\`)`,
+  toLabel: ({ assert }) => `(${ normalizeAssertionVerb( assert.assertion ) } \`${ truncate( assert.value, 60 ) }\`)`,
+  toText: ({ assert }) => `(${ normalizeAssertionVerb( assert.assertion ) } \`${  assert.value, 60 }\`)`,
   commonly: "assert template variable value",
 
   description: `Asserts that variable associated with a given name satisfies the given constraint`,